requirements.txt
annotated-types==0.7.0
anyio==4.6.0
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
httptools==0.6.1
idna==3.10
psycopg==3.2.3
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyMySQL==1.1.1
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
